CacheStorage: Report disk full error correctly when creating a cache.
Past metrics show that the main cause of failure for
base::CreateDirectory in IndexedDB is the device being out of space.
This CL plumbs the error code through so we can properly report this
as a QuotaExceeded exception.

An alternative would be to retry the operation, but if the disk is
truly full then its likely any further operations will run into disk
exhaustion anyway.  For now lets just report the problem as accurately
as we can.

Bug: 1113829
Change-Id: Idc66d2f43b474e1a97229e0057031ba12c96d677
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343305
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795979}