Reduce elide animation gradient mask size by 1
GetSubstringBounds rounds up on both sides (so it guarantees that the
rectangle returned will be >= than the substring). this causes the
ElideAnimation gradient mask to sometimes cover the unelided section of
the URL by one pixel. This CL adds an extra pixel of margin to ensure
this never happens. Tests are also modified to reflect the new
expectations.

Bug: 1112848
Change-Id: Ic120f74f276f1baffdca9509784b2700b7649175
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2341397
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796010}