Fix a pinch zoom laggy problem
This CL fixes a problem when pinch zoom is laggy. The root cause
of this is due to a legacy IPC channel that is not removed yet
in the InputRouterImpl.

More specifically, in InputRouterImpl, we now have:
1. OnSetCompositorTouchAction, which sets the allowed touch action
computed from CC. This is the most restricted touch action which
can be used to process gesture event right away.
2. SetTouchActionFromMain: this is the touch action computed by
the Blink main thread, and that's the correct one (more correct
than the compositor computed one), and we should use this
touch action to process touch events.
3. OnSetTouchAction: this is the legacy one, it was there before
SetTouchActionFromMain was introduced. This should be removed.

The problem is that we got SetTouchActionFromMain to be Auto,
which means any gesture event is allowed. But then OnSetTouchAction
somehow set the touch action to be none, which means no gesture
event can be allowed, and that prevents pinch zoom.

Bug: 1059125, 1030234
Change-Id: I4bb66ad52f6a1353f8383d6685cb293b41d4e08e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2304161
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796042}