overview: Simplify bounds calculations.
The overview bounds calculations today consists of determining the
bounds in OverviewGrid, then adding a margin (for the highlight
border) and to eat events near the edge), then removing the margin for
calculations down the line. This patch simplifies some of that logic,
allowing us to remove a workaround in WmHighlightItemBorder.

Some calculations are still left behind, but its a lot fewer and
better commented.

This does affect behavior in one case - for extreme window cases (in
this example super wide windows), the ratio between the width and height
is now 2:1. Previously, the ratio between the width + margin and height
+ margin was 2:1. I think this is the expected behavior we want.

Test: existing tests
Bug: 1111961
Change-Id: I564a63cdf5f9420ef364cd6473acaccba3c6050a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335794
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796046}