Prepare RenderTests for name changes
Prepares RenderTestRule for the upcoming test name change to not include
OS or device model in the test name reported to Skia Gold. This is done
by creating a copy of the compareForResult method that does not add the
extra information and calling it at the end of compareForResult.

These extra comparisons have the ignore=1 field in their JSON, which
means that they won't cause Gold to comment on CLs if they fail. This is
so that any existing flakiness in the tests can work through the system
with the new names without affecting anyone.

Once everything is ready, this new implementation can just replace the
existing compareForResult (with the ignore field removed and some
renaming performed).

Bug: 1077274
Change-Id: I714999a382b2f2f9a7e4e5abe34d46a817c8882a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337101
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796052}