[conversions] Wrap ConversionStorage in a thread safe context
Moves ConversionStorage and it's task runner into a wrapper which can
be shared across threads. The new ConversionStorageContext is
created by the ConversionManager on the UI thread.

To integrate with QuotaManager, ConversionStorage needs to be
accessible/owned by the IO thread where QuotaClients are accessed.

ConversionStorageContext wraps a ConversionStorage instance which lives
on its own task runner. As is, this requires redeclaring all
ConversionStorage methods with appropriate callbacks on the
ConversionStorageContext. Future improvements to base::SequenceBound
may remove this need, but for now this simply moves the existing
task runner + storage interface into the storage context

This construction allows for data races on tasks that are posted to
the storage task runner, which is okay because operations on
ConversionStorage does not require any sequencing on operations.

Change-Id: Iff0c2c22bb1d02d98f8d8fd80734f235b116918c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239676
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796070}