Views: Create a NewTabButton subclass for the TabSearchButton
This CL creates the initial NewTabButton subclass for the
TabSearchButton. While the Tab Search experiment is running this is
intended to be a transitionary state. The aim is to eventually have
both the NewTabButton and TabSearchButton inherit from a common base
class and specialize where appropriate.

Follow up CLs will add further tweaks including locking the button
highlight while the async button loads invisibly in the background.

Bug: 1099917
Change-Id: Ib1c087c90ef4738500ed28583049ba53bda897c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343384
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796125}