wm: Fix regression with dragging windows with transforms.
In crrev.com/c/2238652, we modified the drag window copy to share codes
with WindowMirrorView. For this case, we skipped setting a transform in
Layout but we actually need to set it to identity since mirroring
layers will copy transforms (case for dragging overview windows). Copies
some more layer properties that may be on the original window.

Test: added test, manual
Bug: 1113429
Change-Id: I7874d733be5026c0785c32cc5ceafb529653b4cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340791
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796139}