PDF Viewer: Highlight download button when menu is open, match bookmarks
The download button should be highlighted when the action menu is open,
similar to the highlight on the bookmarks icon when the bookmarks menu
is open.

Also changing the menu to be light when the PDFViewerUpdate feature
flag is not enabled, for the same reason (matching the bookmarks menu).

Bug: 1113519
Change-Id: I86dc1b7a569688532a2bbce1b8b639e79932d325
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340447
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796144}