Files: Allow path components break-line anywhere in metadata box.
The rendering artifact mentioned in issue 1095864 is actually a
horizontal scrollbar. It is shown unexpectedly when a single path
component in the file location path is longer than the width of
metadata box.
Each path component's CSS overflow-wrap property is break-word (derived),
but it should be OK for path component to break line in the middle of a
word.

Bug: 1095864
Test: Manually tested.
Change-Id: I7e6a986511cf1828957e35fb3549f8c6417b23f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344330
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Naoki Fukino <fukino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796315}