[Autofill] Removed FormStructure::GetServerTypes().
FormStructure::GetServerTypes() created a vector that was always
passed to data_util::DetermineGroups().

This CL removes FormStructure::GetServerTypes() and moves the relevant
computation of the server into the loop in
data_util::DetermineGroups().

We need to keep an overload data_util::DetermineGroups() that takes a
vector because of one reference in LabelFormatter::Create().

Bug: 1007974
Change-Id: Idea5c2d4ea484701ac6a898b4eb6b667859de488
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2266173
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/master@{#796331}