X11 and Ozone: Fix screen initialization for Ozone config
The tests that use ClientHintsBrowserTest override FeatureList
during ::SetUp call, which results in erasing UseOzonePlatform
from that, and InProcessBrowserTest::SetScreenInstance cannot
check the status of the feature flag during ::SetUp call. And,
as a consequence, fail to decide what path it must use. Append
the UseOzonePlatform feature flag back to the scoped feature list
so that initialization happens properly.

-----

PS: Please note that this is a temp solution that will help to choose
between ozone and non-ozone X11 build. The switch that will be used
to choose the path is --enable-features=UseOzonePlatform. Once
non-Ozone X11 path is removed (hopefully by Q1 2021 depending on how
the finch trial goes), the wrapper will be removed.

Please also note that it's impossible to build use_x11 && use_ozone
without some hacks in PlatformCursor code. The changes to that are
on their way to upstream.

----

Bug: 1085700
Change-Id: I995a434a25b6e1a0db3d16bb516184c0d3bc87de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333843
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Maksim Sisov (GMT+3) <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#796341}