X11 and Ozone: Temporarily disable SkiaVulkan backend for PixelTests
When use_x11 && use_ozone are set, vulkan tests start to use
SkiaVulkan backend (that is actually used only for non-Ozone X11
builds on Linux). To suppress the crashes, default to a default
backend for Ozone. Also, added a bug to enable that for Ozone/Linux.

Bug: 1085700
Change-Id: Ib8e99caf046dc91e90a954b4665e241ed9c80f39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336083
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov (GMT+3) <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#796348}