[ozone/x11] Enabled the kSelection clipboard
The selection clipboard is available on Linux/X11 but Ozone hides the
platform so it is only known at run time what is it.  Because of that,
Ozone/X11 could not use the selection clipboard.

This CL introduces a new call into the ClipboardHost Mojo interface so
the renderer can query the host for whether the selection clipboard is
available.

Change-Id: I64fc8ca66509c2b12f3d03fb53aad3e986772153
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307234
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796365}