[signin] Remove lingering invalid account when signing out from settings
To fix issues with SAML signout, the signout flow was changed to go
through Gaia (http://crbug.com/1068978). However, this does not work when
the account is already invalid (Gaia signout is a no-op in this case).

As a result, when signing out from an invalid account (i.e. in Sync
paused state), the account was not immediately removed from the device,
and would linger until the next reconcilor run.

This CL explicitly removes the account if it's invalid.

Fixed: 1114646
Change-Id: Ibd7b4466ecb7d0ac6951d6b30814ef2e5c521bb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346245
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796430}