ash: Add password and PIN views to AuthDialogDebugView
The password and PIN views are used to receive password or PIN
from the user. The relationship between the password and PIN views
(both visually and functionally) should be the same as in lock screen.

Because the in-session auth dialog uses white background, the
password and PIN views are set to other colors to be visible.

A follow-up change will query whether the user has PIN set up and hide
the PIN view if there's no PIN.

Bug: b:156258540
Change-Id: I63ab51839ce913ee61f159e3e56861dd9f4f735f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342364
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796512}