Make fieldtrial annotation work with features specified in command line
Before this CL, when FieldTrials.collectFieldTrials() is invoked,
mergeFeatureLists() has not been called, so enabled features are not
correctly populated. This makes collectFieldTrials() produce incorrect
data, which leads to the crash in applyFieldTrials().

This CL makes sure the enabled/disabled features specified in the
command line arguments are visible to FieldTrials.

Bug: 1113847, 1102633
Change-Id: I5c49c41c87fd0219c73a967db2ee33dc8237f52f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344805
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796552}