Stops empty suggestion headers from showing up in the NTP realbox
The server provided suggestion header map may come with entries
that do not correspond to any of the matches. Using that map to display
suggestion headers can result in empty headers. This CL fixes that by
using the header information in the matches themselves as the source of
truth about what headers actually exist in the server response. It also
modifies the type annotations to consistently use the "number" type for
the suggestion IDs.

Bug: 1113447
Change-Id: I5f685595cfe5a295d483447e6965cb77b7c581fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347047
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796560}