Reland r786979, part 2: add hooks to ToolbarAccountIconContainer.
The original CL was landed in one piece and reverted due to apparent
perf regressions.  Reland in sections to try and determine more
precisely whether there is any perf impact.

This CL relands the pieces that will be necessary to use callbacks
instead of ButtonObserver in ToolbarAccountIconContainer, but does not
actually switch to using them yet.

If this causes regressions, then something about the memory overhead of
the subscription list, or the cost of Button::GetHighlighted(), is the
issue.

Bug: none
Change-Id: I34921876c23579c01ca378ad4fc2ae6fee9042fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347312
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796616}