Use given trace file directly if the temp file doesn't work
Fall back to the old behavior of just writing to a trace file directly
if we're unable to open a temporary file in the same directory. This can
happen in some configurations on restrictive Android systems for
example, where the browser may only be able to write to the actual file
specified in --trace-startup-file, and not to any nearby path.

Bug: 1082916
Change-Id: I3c1275b4c70cc8edb6e50adc321ba412c6226de8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346352
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Tomasz Śniatowski <tsniatowski@vewd.com>
Cr-Commit-Position: refs/heads/master@{#796735}