[ozone/x11] Made the platform returning if transparent color would work.
The Textfield::WriteDragDataForView sets transparent background color
for the drag image, which is not always supported on X11 (depends on
whether the compositing manager is available), but Ozone/X11 didn't
check that and always used the transparent color.

This CL adds a new method in the Ozone platform that returns whether
transparent color could be used.

Bug: 1085700
Change-Id: I2c812086993fc825bb8f7bc0d416b7aa4ab44d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2323293
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Maksim Sisov (GMT+3) <msisov@igalia.com>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#796738}