[PwdCheckAndroid] Implement native side of deletion
In order to natively delete a credential, the compromised credential on
the java-side needs to contain unformatted information about username,
origin, sign-on real, and password. Therefore, this CL:
 * adds the required key attributes
 * makes the key attributes accessible by making JNI headers available
 * makes interfaces that pass 7 attributes to construct a credential
   pass the credential directly instead (minimizes confusion)

The actual delete call is only one line in the password_check_manager
that triggers the deletion and indirectly updates to the UI once the
deletion happened.

The deletion promoted two smaller errors which were one-line-fixed:
 * the mediator never clear old credentials — it just appended new ones
 * the confirmation dialog triggered twice (once with a null handler)

Screenshot of the entire flow in the linked bug.

Bug: 1108358, 1092444
Change-Id: I1f1a3f6e73c40c316cb2f711c44236388ff05b4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346345
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Andrey Zaytsev <andzaytsev@google.com>
Cr-Commit-Position: refs/heads/master@{#796821}