Add id to cast web contents observer
Added a id to replace tab_id as unique
identifier for the webviews that are created from the
platform views service.  The tab_id was always 0
because is_root_frame is true on creation.
Subsequent lookups for the ax tree id by tab id
were only working by chance. The linear search for the
tab id among the list of windows always matches the
first one found, which was the correct one.  This ensures
tab ids will be unique.

Bug: None
Test: display assistant created webviews followed by navigation
      causing render frames to be swapped

Change-Id: Ib6b9f1fd217367423f15965e5b760df1f170caa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342640
Commit-Queue: Randy Rossi <rmrossi@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796903}