[Native Cast MRP] Presentation connection for site-initiated mirroring
This fixes a regression relative to the extension MRP where we were
establishing a presentation connection only if a client ID is set.
If mirroring is initiated by a site with cast:0F5096E8 there is no
client ID but we still need a presentation connection.

Bug: 1114891, b/163366468
Change-Id: Ib66768c52851bb87aba307775d917b81bf669572
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347890
Commit-Queue: Takumi Fujimoto <takumif@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#796940}