Use a smaller capture region in Pointer Lock
The capture region that held the mouse inside the chrome window was the
full size of the window. This caused issues while when the invisible
mouse hit the edge of the window it could fire certain events and cause
strange things to happen. For example the linked bug discusses an "x"
appearing when you are also in fullscreen. There are also some dual
monitor setups that have consistently proven that the mouse moves just
out of the window and onto the other monitor while in fullscreen. To
solve this this CL is adding a 5 pixel border to the edge of the capture
region to give some wiggle room for the small Windows errors and to stop
events from firing when the mouse hits the edge of the window.

Bug: 1107507
Change-Id: Ia2314f7726b5a18d96800a9efcd990bace564ef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343005
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797072}