Allow using an 'optional' font as long as it loads before first use
After launching crbug.com/1040632, we disallow a 'font-display:optional'
font from being used if it doesn't load before the first rendering cycle
begins, so that it won't cause any relayout.

This turns out to be too stringent. This patches allows using the font
as long as we have never rendered fallback for it. In this way, we still
guarantee no layout shifting, and the font can be used in more
circumstances.

Bug: 1114314
Change-Id: Id521e97b8882e2e3d1fb13c09e040c8e2fd3d47e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343643
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797076}