Fixes CharacterWidths() and text Len(), and simplifies OffsetInContainer() in AbstractInlineTextBox
The text length of an abstract inline text box was sometimes reported as one character short because
the trailing white space was not included, whilst it was included in the actual text.
Similarly, the character widths did not include the width of any trailing white space.
Also simplified the calculation of OffsetInContainer by using some existing helper methods in Layout NG
and improved tests.

AX-Relnotes: n/a.

R=dmazzoni@chromium.org, aleventhal@chromium.org

Change-Id: Ia5af8a475c028ecc1980b7cbe12695807164824d
Bug: 1008031
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346289
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797185}