[iOS] Support scribble in fakebox.
Adds support for scribble in fakebox. For this, an indirect scribble
interaction is added to the fakebox. This interaction delays focusing,
which allows users to write without the fakebox jumping away. Then it
forwards the scribble events to the omnibox.
Adds a ton of plumbing to connect the fakebox and the omnibox.

Bug: 1098342
Change-Id: I3a5783218b5a25dc8146f53252e50348ac1886d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340978
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797211}