Rename *MoveToAccountRefusedCount to *MoveOfferedToNonOptedInUserCount
Turns out it's easier to record when the move flow is offered rather
than when it's refused/ignored. For non-opted-in users these things are
basically the same, since accepting the move bubble prompts an opt-in.
No behavior is changed, the counter still isn't incremented.

Bug: 1082152
Change-Id: I379e137d8c24e02625d569d9b65ddd715050ae58
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352229
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#797227}