[iOS CBCM] Support enrollment and policy fetch
Implements the missing delegate methods for enrollment and policy
fetch. Note that there is still an issue with the specific
URLLoaderFactory implementation handling the requests that prevents
the end-to-end flow from working properly. This is being investigated
separately.

Also adds CloudManagementEnrollmentToken to the list of supported
policies (even though it's read directly from NSUserDefaults).

#Squeegee

Bug: 1066495, 1111145
Change-Id: I21c1ee3b26bc413ff26ad4031b44ebda5086e936
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337267
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797274}