Rewrote PRESUBMIT.py with new approach.
Re-design of the CL (crrev.com/c/2341362) to enable presubmit_support
to time/ResultDB wrap each individual lower-level check in PRESUBMIT.py.
The approach taken here is to have the PRESUBMIT_VERSION = 2.0.0 in here
so that presubmit_support is aware of the new version of PRESUBMIT.py,
then check input_api.version in CheckChangeOnCommit/Upload to ensure
that presubmit_support itself is the new version.

The associated changes to depot_tools (crrev.com/c/2349979) must be
landed in order for this to work.

Change-Id: I91d09337dd77a74540ac4a01017275bcaaa67cde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350234
Commit-Queue: Saagar Sanghavi <saagarsanghavi@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#797291}