[WebLayer] Fix URL bar appearance in Weblayer Shell.
The appearance of the URL bar in Weblayer shell was broken in:
https://chromium-review.googlesource.com/c/chromium/src/+/2330614

when we renamed menu_button to app_menu_button but didn't update the reference
to it in the layout. This fixes that and restores the initial size of
the URL bar as well as the horizontal centering.

Before this change:
https://drive.google.com/file/d/1_m1hbatI6D-Nd988RoBnjv28_Y77rJRp/view?usp=sharing
After this change:
https://drive.google.com/file/d/1UBThgpxFDbgg_j93MGkwhQQlIYDz8WnU/view?usp=sharing

Bug: 1114238
Change-Id: I2abbf85a188610d0b334558112b0397243ada4fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352729
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Auto-Submit: Mugdha Lakhani <nator@chromium.org>
Reviewed-by: Robbie McElrath <rmcelrath@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797295}