Rephrase comments in ClientVariations proto
As part of a DevTools CL [1], we expose decoded X-Client-Data HTTP
header values as code in a format similar to the proto definition.
Since we want to include comments in that user-visible output, there
is a desire to match the comments in the source.

This patch tweaks the comments in the source to align with the desired
UI strings (slightly shorter, and using active voice).

[1]: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2351775

No-Presubmit: true
Bug: chromium:1103854
Change-Id: Ic128e0289559ab95c6e293dbcd15aead7cebe6cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351928
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797312}