[Android][Signin] Move key used to store email to ChromeBackupAgent
Even though ChromeBackupAgent no longer gets the sign-in account from
ChromeSigninController, it still uses
ChromeSigninController.SIGNED_IN_ACCOUNT_KEY as the key to store the
email of the signed in account. This CL moves this key to
ChromeBackupAgent. The value of the key is preserved for backward
compatibility.

Bug: 1046412
Change-Id: I4c49a406f474348471de2f27c67c6e4ac2c523f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339457
Reviewed-by: Alice Wang <aliceywang@chromium.org>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797324}