CrOS: Ensure maximum size on standardized icons that are already circular
This change ensures that circular icons are scaled down when needed so
that they do not exceed a maximum size relative to the size of the
bitmap.

Before this change, when circular icons were standardized they are
returned without any changes, even when the size of the icon is too big.

Also in this change is an additional unittest, to ensure that a large
circular icon is scaled down to the maximum size when standardized.

Bug: 1110496
Change-Id: I032da9422d078db8ae8e5d1d1198c8e12f3dec76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347318
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797328}